fixed compiler error under vs2008
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@2792 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cv/cvstereobm.cpp
@@ -862,7 +862,7 @@ static void findStereoCorrespondenceBM( const Mat& left0, const Mat& right0, Mat
 #else
     N0 /= 4;
 #endif
-    double maxStripeSize = min(max(N0 / (width * ndisp), (wsz-1) * SAD_overhead_coeff), height);
+    double maxStripeSize = min(max(N0 / (width * ndisp), (wsz-1) * SAD_overhead_coeff), (double)height);
     int nstripes = cvCeil(height / maxStripeSize);
 #else
     const int nstripes = 1;
